first organizing push
made new main folders for auto, teleop, and commons
